# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RESTful API for a simple social media application. It allows users to manage their posts, including creating, updating, and deleting posts, and provides authentication using JWT. The API is built using the Fiber framework and interacts with a PostgreSQL database.",
        "title": "RAION ASSESSMENT API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8084/api/v1",
    "basePath": "/",
    "paths": {
...
        "/posts/user/{user_id}": {
            "get": {
                "description": "Get all posts made by a specific user, including the caption, image URL, and timestamps.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts by a specific user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful fetch posts by user response",
                        "schema": {
                            "$ref": "#/definitions/response.GetAllPostsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/search/users": {
            "get": {
                "description": "Search for users by their name or email.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful search response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.SearchUsersResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retrieve a list of all users from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "Successful fetch users response",
                        "schema": {
                            "$ref": "#/definitions/response.GetAllUsersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the bio, image_url, and/or username of the authenticated user.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Updated username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Updated bio (optional)",
                        "name": "bio",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Updated image (optional)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful update user response",
                        "schema": {
                            "$ref": "#/definitions/response.UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrieve the details of a specific user by their ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful fetch user by ID response",
                        "schema": {
                            "$ref": "#/definitions/response.GetUserByIDResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.GetAllPostsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Post"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.GetAllUsersResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.User"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetUserByIDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/response.User"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.Post": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "Had an amazing day at the beach!"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-31T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "f9d6b52a-76a1-4b2b-9229-4c8db23a5ef2"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/images/beach.jpg"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-31T12:30:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "2e0850c7-d213-4a91-9b78-bb86e3a6f0d3"
                }
            }
        },
...
        "response.SearchUsersResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.User"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/response.User"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.User": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Hi there!"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-31T12:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "3d5a8b92-f1c5-4dbe-a2a7-1d9a8c743e9b"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/profile.jpg"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-31T12:30:00Z"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}